requirements.txt
requests
beautifulsoup4
pandas
feedparser
pyyaml
matplotlib
